requirements.txt
gspread
google-auth
streamlit
uuid
pandas
path
typing
numpy